refactor(imports): migrate deep imports to barrel — models + types
Create models/index.ts and types/index.ts barrel re-export files.

Migrate 31 source-file deep imports to barrel paths:
- models/roles -> models (20 imports)
- models/GameStatus -> models (10 imports)
- models/Template -> models (1 import)
- types/RoleRevealAnimation -> types (5 imports merged)

Merge multi-source imports into single barrel statements where
files previously imported from both models/roles and
models/GameStatus. Spec-internal deep paths (models/roles/spec/*)
consolidated to spec barrel where possible.
