+++ packages/game-engine/src/engine/state/buildInitialState.ts
@@ -6,8 +6,7 @@
  * 纯函数，无副作用，不依赖 React Native / Expo / IO。
  */
 
-import { GameStatus } from '../../models/GameStatus';
-import type { GameTemplate } from '../../models/Template';
+import { GameStatus, type GameTemplate } from '../../models';
 import type { GameState } from '../../protocol/types';
 
 export function buildInitialGameState(
