+++ packages/game-engine/src/models/index.ts
@@ -1,0 +1,10 @@
+/**
+ * Models barrel — re-exports all model modules.
+ *
+ * Internal game-engine code should import from this barrel (`../models`)
+ * rather than deep paths (`../models/roles`, `../models/GameStatus`).
+ */
+
+export * from './GameStatus';
+export * from './roles';
+export * from './Template';
