+++ packages/game-engine/src/types/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Types barrel — re-exports all type modules.
+ *
+ * Internal game-engine code should import from this barrel (`../types`)
+ * rather than deep paths (`../types/RoleRevealAnimation`).
+ */
+
+export * from './RoleRevealAnimation';
